+++ CHANGELOG.md
@@ -28,6 +28,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added string error code to the Subscribe APIs to differentiate between different 404 scenarios. remp/web#2263 
 - Fix Mailer segment provider users acquiring. Provided segment code needs to be processed before fetching users from database. remp/mnt#114
 - Fix New template generator form - broken sorting value `after`. If selected, select box was not shown. remp/helpdesk#2073
+- Added command `crm:validate-emails` to validate all email addresses for users in a given time period. remp/remp#1026
+  - You can enable this command in your config.neon if you already defined `crmClient` service:
+  ```
+  services:
+    console:
+        setup:
+            # Enable only if "crmClient" service is available
+            - add(Remp\MailerModule\Commands\ValidateCrmEmailsCommand())
+  ```
 
 ## Archive